backend/requirements/development.txt
# Include base requirements
-r base.txt

# Development and debugging tools
django-debug-toolbar==4.2.0
django-extensions==3.2.3

# Code quality and formatting
black==23.9.1
flake8==6.1.0
isort==5.12.0

# Testing
pytest==7.4.2
pytest-django==4.5.2
pytest-cov==4.1.0
factory-boy==3.3.0

# Documentation
Sphinx==7.2.6
sphinx-rtd-theme==1.3.0